refactor： - sample deletion email summary - directory optimization
fix：
- the default value should be 'chatgptapi' instead of 'chatgpt'
